Address @afeld's comments from #24
