+++ lib/jekyll_pages_api/page.rb
@@ -3,7 +3,7 @@ require_relative 'filters'
 module JekyllPagesApi
   # wrapper for a Jekyll::Page
   class Page
-    HTML_EXTS = %w(.html .md .markdown .textile)
+    HTML_EXTS = %w(.html .md .markdown .textile).to_set
     attr_reader :page
 
     def initialize(page, site)
@@ -12,7 +12,7 @@ module JekyllPagesApi
     end
 
     def html?
-      path = @page.path
+      path = self.page.path
       path.end_with?('/') || HTML_EXTS.include?(File.extname(path))
     end
 
@@ -22,7 +22,7 @@ module JekyllPagesApi
 
     def title
       title = self.page.data['title'] if self.page.respond_to?(:data)
-      title = self.page.title if title == nil and self.page.respond_to?(:title)
+      title ||= self.page.title if self.page.respond_to?(:title)
       self.filterer.decode_html(title || '')
     end
 
@@ -30,17 +30,19 @@ module JekyllPagesApi
       @site.baseurl
     end
 
+    def rel_path
+      path = self.page.url if self.page.respond_to?(:url)
+      path ||= self.page.relative_path if self.page.respond_to?(:relative_path)
+      path
+    end
+
     def url
-      rel_path = self.page.url if self.page.respond_to?(:url)
-      if rel_path == nil and self.page.respond_to?(:relative_path)
-        rel_path = self.page.relative_path
-      end
       [self.base_url, rel_path].join
     end
 
     def body_text
       output = self.page.content if self.page.respond_to?(:content)
-      File.open(self.page.path, 'r') {|f| output = f.read} if output == nil
+      output ||= File.read(self.page.path)
       self.filterer.text_only(output)
     end
