+++ lib/jekyll_pages_api/version.rb
@@ -1,3 +1,3 @@
 module JekyllPagesApi
-  VERSION = '0.1.3'
+  VERSION = '0.1.2'
 end
